restructuring code and adding path aliases
