+++ app/components/game/tic-tac/tic-tac-cell/reset/Reset.ts
@@ -1,7 +1,7 @@
-import {getTd, getTdCell, tdClassList, setUnClicked, resetStopAnimateMoveSuccess} from '@tic-tac/tic-tac-cell/Common';
+import {getTd, getTdCell, tdClassList, setUnClicked, resetStopAnimateMoveSuccess} from '@tic-tac/tic-tac-cell/common';
 import {ChangeFunctionType, ColumnIdType, TDClassIdType} from '@types-dir/index';
 import {InitializeContextsFunctionType} from '@contexts/index';
-import {addClickListener, removeClickListener} from '@tic-tac/tic-tac-cell/Onclick/OnClick';
+import {addClickListener, removeClickListener} from '@tic-tac/tic-tac-cell/on-click/OnClick';
 
 const removeAllNewClasses = (
   columnId: ColumnIdType,
